Make deployment ui display single container
In order to reduce clutter, deployment service now only displays single
container. In order for this to work, we added another preference to the
deployment plug-in.

Along the way, separated plug-in completely from DICE Platform, since there is
no reason for such tight coupling between the two of them. In fact, this
coupling made adding new preference to plug-in quite painful.

Additionally, with new separation, this plug-in is usable on it's own (without
having complete DICE Platform installed). This means that it can be used to
deploy hand-written blueprints directly from Eclipse.