Add user interface to plugin
This commit adds three user interface elements that are used to
configure plugin and monitor deployments.

First part is preferences page where user can specify services that
can be used to deploy blueprints. Information about services is stored
by datastore backend in a secure manner, so user will need to provide
password when interacting with data store for the first time.

Second part is run dialog that makes it possible to actually execute
deployment using user-selected blueprint. There is some information on
this dialog present that is not currently used by plugin, but will be
needed as soon as we try to deploy any real-world blueprint with
bundled resources.

Third part is simple container view that displays all containers that
are available to the user and their status.