Force container update before deploy is started
This ensures that container store is populated with containers that
are available on DICE Deployment Service at the moment of the run.

Another scenario that would cause deploy to fail is if
ContainerProvider is used for the first time in deployment delegate.
Because store updates are done asynchronously in the background,
deployment delegate would get NULL as a result of the container query.