Add DICE platform compatibility plugin
Because DICE platform does not offer common UI elements as a separate
feature that we would be able to depend on, we created a shim that
gives us the missing functionality.

This shim makes it possible to have stand-alone installation of plugin
without having to get the complete DICE IDE. Just include extra
feature when installing plugin and any missing UI elements will be
added.