Expose blueprint packaging errors to user
Since there is not much we can do when packaging fails, error dialog
is presented to user and deployment is terminated prematurely.

Maybe adding a full stack trace to the error is a bit of an overkill,
but since stack trace is hidden by default, there should be no
inconveniences for end users. But this stack trace may prove itself
pretty useful when user reports a bug.