Unify DICE HTTP client exceptions
Previously, client was throwing mix of IOExceptions and ClientErrors
in its public methods, which was not optimal and complicated error
handling in code that uses DICE client.

To remedy this situation, we consolidated all exceptions into
ClientError, which is now the only exception that is being thrown in
client. Code that uses client now only needs to handle single
exception, which should make try/catch clauses a bit cleaner.