Implement proper error handling in deploy driver
Now all of the errors that happen during deployment process should be
reported to the user.

Tasks are now aborted cleanly by raising core exception that notifies
Eclipse that task failed to terminate properly. This should get rid of
the last few error messages that eclipse regurgitated if we terminated
it after failed run attempt.