Increase default socket timeout in HTTP client
When deploying blueprints to FCO, default socket timeout of 5 seconds
proved itself to be to too short when platform is under heavy load.

In order to fix this, we increased the default timeout to one minute,
which should be more than enough, since most of the server operations
are asynchronous and response is returned as soon as possible.