Add service validation to launch configuration
Simply presenting an empty service selection combo box to the user
without any hints about what actions would be appropriate to take is
not something we should do. This is why we added another validation to
the launch configuration dialog that checks if user has any services
registered with the plugin. If there is no services available, user is
advised to add new service in plugin preferences.