Make it possible to send arbitrary data to service
Client is now capable of sending arbitrary data to deployment service.
This makes it possible to send blueprint metadata at deployment time.

Note that client does not enforce any rules upon the data being sent.
It is responsibility of the user to provide proper content and handle
errors on malformed data.