Prevent storing stale values in launch dialog
Current implementation of service/container selection in run dialog
had troubles keeping values updated and in sync with values stored in
run configuration. Main cause of this behavior is asynchronous nature
of service updates that are required in order to keep UI responsive.

Simplifying update logic removed all of those troubles at the cost of
updating values even if this is not strictly needed, but since user is
already clicking around our dialog, this additional delay should not
be noticeable.