Add scripting for splitting update sites
The script `prepare-release.sh` now takes care of creating an
official updates site `updates/`. First it builds the full plug-in
with all the needed dependencies in the `updates_standalone/`. Then
it copies the ide-specific stuff to the `updates/VERSION` folder.
It also takes care of the artifacts and contents xmls to account
for all the available plug-in versions.