Limit blueprint extensions by default
Blueprints have no designated file extension since tools that are
processing them only expect that blueprint file contains valid yaml.
But in order to reduce the amount of files users see in the file
selection dialog, we added a default filter that limits display to
*.yaml, *.yml and *.txt files.

In order to make it possible to use arbitrary extension for blueprint
files, we also added *.* filter that mimics previous behavior.